Refactor SSOStateStoreManager to different location.
